requirements.txt
json5==0.9.5
python-telegram==0.12.0
redis==3.5.3
termcolor==1.1.0